1303/quantum-gpu/requirements.txt
# =============================================================
#  Quantum-Classical Hybrid Neural Network — Dependencies
# =============================================================
#
#  IMPORTANT (Windows + NVIDIA GPU):
#  Install PyTorch with CUDA *before* the rest of the deps:
#
#    pip install torch --index-url https://download.pytorch.org/whl/cu121
#
#  Then install everything else:
#
#    pip install -r requirements.txt
#

# Core ML
# NOTE: torch is expected to be pre-installed with CUDA support (see above).
#       If not yet installed, the line below pulls CPU-only as fallback.
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Quantum Computing
pennylane>=0.35.0
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-machine-learning>=0.7.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration & CLI
pyyaml>=6.0
click>=8.1.0

# Logging
tensorboard>=2.15.0

# Notebooks
ipykernel>=6.25.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pandas>=2.1.0
tabulate>=0.9.0
